Start refactoring backend consumer code to support multiple topics
